+++ apps/demo/src/components/header.tsx
@@ -1,22 +1,33 @@
 'use client';
 
 import Link from "next/link";
-import { useDesign, UserButton } from "@stackframe/stack";
+import { useDesign, UserButton, ColorPalette } from "@stackframe/stack";
+import { useTheme } from "next-themes";
+import styled from "styled-components";
 import ColorMode from "./color-mode";
 import Select from "./select";
 import { useCurrentUI } from "./provider";
 
+const StyledHeader = styled.div<{ $colors: ColorPalette }>`
+  border-bottom: 1px solid ${props => props.$colors.light.neutralColor};
+  background-color: ${props => props.$colors.light.backgroundColor};
+
+  html[data-theme='dark'] & {
+    border-color: ${props => props.$colors.dark.neutralColor};
+    background-color: ${props => props.$colors.dark.backgroundColor};
+  }
+`;
+
 export default function Header() {
-  const { colors, colorMode } = useDesign();
-  const [currentUI, setCurrentUI] = useCurrentUI();
+  const { theme, setTheme } = useTheme();
+  const { colors } = useDesign();
+  const { ui, setUi } = useCurrentUI();
+
   return (
     <>
-      <div 
+      <StyledHeader
+        $colors={colors}
         className={"fixed w-full z-50 p-4 h-12 flex items-center py-4 border-b justify-between"}
-        style={{
-          borderColor: colors.neutralColor,
-          backgroundColor: colors.backgroundColor,
-        }}
       >
         <Link href="/" className="font-semibold">
           Stack Demo
@@ -28,13 +39,13 @@ export default function Header() {
               { value: 'default', label: 'Default UI' },
               { value: 'joy', label: 'Joy UI' }
             ]}
-            value={currentUI}
-            onChange={(e) => setCurrentUI(e.target.value as any)}
+            value={ui}
+            onChange={(e) => setUi(e.target.value as any)}
           />
           <ColorMode />
-          <UserButton />
+          <UserButton colorModeToggle={() => setTheme(theme === 'dark' ? 'light' : 'dark')} />
         </div>
-      </div>
+      </StyledHeader>
       <div className="min-h-12"/> {/* Placeholder for fixed header */}
     </>
   );