+++ packages/stack/src/components-core/input.tsx
@@ -1,19 +1,15 @@
 'use client';
 
 import React from "react";
-import { useDesign } from "../providers/design-provider";
+import { ColorPalette, useDesign } from "../providers/design-provider";
 import styled from 'styled-components';
 import { BORDER_RADIUS, FONT_FAMILY, FONT_SIZES, PRIMARY_FONT_COLORS, SECONDARY_FONT_COLORS } from "../utils/constants";
 
 type InputProps = React.InputHTMLAttributes<HTMLInputElement> & Pick<React.HTMLProps<HTMLInputElement>, 'ref'>
 
+
 const StyledInput = styled.input<{
-  $borderColor: string,
-  $focusShadowColor: string,
-  $textColor: string,
-  $placeholderColor: string,
-  $fileSelectorColor: string,
-  $fileSelectorButtonColor: string,
+  $colors: ColorPalette,
 }>`
   font-family: ${FONT_FAMILY};
   font-size: ${FONT_SIZES.md};
@@ -21,45 +17,61 @@ const StyledInput = styled.input<{
   border-radius: ${BORDER_RADIUS};
   background-color: transparent;
   border: 1px solid;
-  border-color: ${props => props.$borderColor};
   padding: 0rem 1rem;
-  color: ${props => props.$textColor};
-  &::placeholder {
-    color: ${props => props.$placeholderColor};
-  }
-  &:focus-visible {
-    outline: none;
-    box-shadow: 0 0 0 2px ${props => props.$focusShadowColor};
-  }
   &:disabled {
     cursor: auto;
     opacity: 0.5;
   }
-  &[type=file] {
-    color: ${props => props.$fileSelectorColor};
-  }
   &[type=file]::file-selector-button{
     border: none;
-    color: ${props => props.$fileSelectorButtonColor};
     background-color: transparent;
     height: 2.5rem;
     margin-right: 0.5rem;
     padding: 0;
   }
+
+  border-color: ${props => props.$colors.light.neutralColor};
+  color: ${PRIMARY_FONT_COLORS.light};
+  &::placeholder {
+    color: ${SECONDARY_FONT_COLORS.light};
+  }
+  &:focus-visible {
+    outline: none;
+    box-shadow: 0 0 0 2px ${props => props.$colors.light.primaryColor};
+  }
+  &[type=file] {
+    color: ${SECONDARY_FONT_COLORS.light};
+  }
+  &[type=file]::file-selector-button{
+    color: ${PRIMARY_FONT_COLORS.light};
+  }
+
+  html[data-theme='dark'] & {
+    border-color: ${props => props.$colors.dark.neutralColor};
+    color: ${PRIMARY_FONT_COLORS.dark};
+    &::placeholder {
+      color: ${SECONDARY_FONT_COLORS.dark};
+    }
+    &:focus-visible {
+      outline: none;
+      box-shadow: 0 0 0 2px ${props => props.$colors.dark.primaryColor};
+    }
+    &[type=file] {
+      color: ${SECONDARY_FONT_COLORS.dark};
+    }
+    &[type=file]::file-selector-button{
+      color: ${PRIMARY_FONT_COLORS.dark};
+    }
+  }
 `;
 
 const Input = React.forwardRef<HTMLInputElement, InputProps>(
   (props, ref) => {
-    const { colors, colorMode } = useDesign();
+    const { colors } = useDesign();
     return (
       <StyledInput
         ref={ref}
-        $borderColor={colors.neutralColor}
-        $placeholderColor={SECONDARY_FONT_COLORS[colorMode]}
-        $textColor={PRIMARY_FONT_COLORS[colorMode]}
-        $fileSelectorColor={SECONDARY_FONT_COLORS[colorMode]}
-        $fileSelectorButtonColor={PRIMARY_FONT_COLORS[colorMode]}
-        $focusShadowColor={colors.primaryColor}
+        $colors={colors}
         {...props}
       />
     );