+++ packages/stack/src/index.tsx
@@ -20,9 +20,11 @@ export { default as UserButton } from "./components/user-button";
 export { default as AccountSettings } from "./components-page/account-settings";
 
 export { useDesign } from './providers/design-provider';
+export type { ColorPalette } from './providers/design-provider';
 export { useComponents } from './providers/component-provider';
-export { StackTheme, ThemeConfig } from './providers/theme-provider';
+export { StackTheme } from './providers/theme-provider';
+export type { ThemeConfig } from './providers/theme-provider';
 
-export { CurrentUser, Project, ServerUser, ApiKeySetFirstView, ApiKeySet } from './lib/stack-app';
+export type { CurrentUser, Project, ServerUser, ApiKeySetFirstView, ApiKeySet } from './lib/stack-app';
 
 export * from './components-core';