+++ docs/CHANGELOG.md
@@ -1,5 +1,23 @@
 # stack-docs
 
+## 0.0.5
+
+### Patch Changes
+
+- CommonJS support
+
+## 0.0.4
+
+### Patch Changes
+
+- Partial pre-rendering
+
+## 0.0.3
+
+### Patch Changes
+
+- E2E tests and various fixes
+
 ## 0.0.2
 
 ### Patch Changes