+++ packages/init-stack/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @stackframe/init-stack
 
+## 1.0.3
+
+### Patch Changes
+
+- Partial pre-rendering
+
+## 1.0.2
+
+### Patch Changes
+
+- E2E tests and various fixes
+
 ## 1.0.1
 
 ### Patch Changes